some CGFloat conversions
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/tests/examples/trunk@36628 72102866-910b-0410-8b05-ffd578937521

+++ gui/GFractal/FractalView.m
@@ -236,7 +236,7 @@ static id cschemes[] =
   int rept, impt;
   double zzbar, esc;
   
-  float colr, colg, colb;
+  CGFloat colr, colg, colb;
   
   id colors, color;
   int count, maxit, cols;
